Working chatbot integrated with API.
Update-Trained data with new tags
